% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RuleInduction.OtherFuncCollections.R
\name{addDescriptorCN2}
\alias{addDescriptorCN2}
\title{Adding a descriptor to a rule in CN2 algorithm}
\usage{
addDescriptorCN2(rulesList, descCandidates)
}
\description{
Adding a descriptor to a rule in CN2 algorithm
}

