% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RuleInduction.OtherFuncCollections.R
\name{bestFirst}
\alias{bestFirst}
\title{the best-first voting strategy function}
\usage{
bestFirst(object, ruleValues, consequents, majorityCls, ...)
}
\arguments{
\item{object}{a class "RuleSetRST".}

\item{ruleValues}{values of rules.}

\item{consequents}{values on the consequent part.}

\item{majorityCls}{a value representing majority class of decision attribute.}

\item{...}{other parameters.}
}
\value{
predicted values.
}
\description{
An auxiliary function that performs the best-first voting strategy.
}

