% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{boolean.func}
\alias{boolean.func}
\title{it is used to perform boolean function}
\usage{
boolean.func(disc.list)
}
\arguments{
\item{disc.list}{a list of attributes resulted by build.discMatrix.FRST}
}
\description{
it is used to perform boolean function
}

