% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{calc.all.reducts}
\alias{calc.all.reducts}
\title{This function is used to calculate a discernibility function used to get all reducts}
\usage{
calc.all.reducts(discernibilityMatrix)
}
\arguments{
\item{discernibilityMatrix}{an object of a class "DiscernibilityMatrix}
}
\description{
This function is used to calculate a discernibility function used to get all reducts
}

