% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{calc.conorm}
\alias{calc.conorm}
\title{conorm calculation}
\usage{
calc.conorm(func.conorm, data, init.val, t.conorm = "lukasiewicz", ...)
}
\arguments{
\item{func.conorm}{a function to calculate 2 variable using a certain conorm}

\item{data}{a vector of data}

\item{init.val}{a value of attribute}

\item{t.conorm}{a type of t-conorm}
}
\description{
This function is used to calculate conorm all relations
}

