% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NearestNeigbour.OtherFuncCollections.R
\name{calc.membershipClass}
\alias{calc.membershipClass}
\title{This function is to calculate membership class of nearest neighbor}
\usage{
calc.membershipClass(nearest.dt, num.class, type = "gradual")
}
\arguments{
\item{nearest.dt}{a matrix of data of nearest neighbor}

\item{num.class}{the number of class}

\item{control}{a list of parameters}
}
\description{
This function is to calculate membership class of nearest neighbor
}

