% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{calc.SAT}
\alias{calc.SAT}
\title{this function is used to calculate degree of satisfaction}
\usage{
calc.SAT(IND.cond, IND.dec, attributes)
}
\arguments{
\item{IND.cond}{indiscernibility matrix for all conditional attributes}

\item{IND.dec}{indiscernibility matrix for decision attribute}

\item{attributes}{a list of considered attributes}
}
\description{
this function is used to calculate degree of satisfaction
}

