% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NearestNeigbour.OtherFuncCollections.R
\name{calc.similarityDegree}
\alias{calc.similarityDegree}
\title{This function is to calculate degree of similarity}
\usage{
calc.similarityDegree(nearest.dt, control)
}
\arguments{
\item{nearest.dt}{a matrix of data of nearest neighbor}

\item{control}{a list of parameters}
}
\description{
This function is to calculate degree of similarity
}

