% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RuleInduction.OtherFuncCollections.R
\name{computeAQcovering}
\alias{computeAQcovering}
\title{Computation of a covering of a lower approximation of a concept by decision rules using the AQ algorithm.
It is an auxiliary function.}
\usage{
computeAQcovering(concept, attributeValuePairs, dataTab, epsilon = 0.05,
  K = 2)
}
\description{
Computation of a covering of a lower approximation of a concept by decision rules using the AQ algorithm.
It is an auxiliary function.
}

