% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{computeCore}
\alias{computeCore}
\title{a function for computing a core from a list of all reducts of a data set}
\usage{
computeCore(reductList)
}
\description{
a function for computing a core from a list of all reducts of a data set
}

