% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RuleInduction.OtherFuncCollections.R
\name{computeLEM2covering}
\alias{computeLEM2covering}
\title{Computation of a covering of a lower approximation of a concept by decision rules using the LEM2 algorithm.
It is an auxiliary function.}
\usage{
computeLEM2covering(concept, attributeValuePairs, decisionValues, uniqueCls)
}
\description{
Computation of a covering of a lower approximation of a concept by decision rules using the LEM2 algorithm.
It is an auxiliary function.
}

