% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{computeRelevanceProb}
\alias{computeRelevanceProb}
\title{An auxiliary function for computing attribute relevance using random probes.
It is used by FS.DAAR.heuristic.RST function.}
\usage{
computeRelevanceProb(INDclasses, attributeVec, attrScore, decisionVec,
  baseChaos, qualityF = X.gini, nOfProbes = 100, withinINDclasses = FALSE)
}
\description{
An auxiliary function for computing attribute relevance using random probes.
It is used by FS.DAAR.heuristic.RST function.
}

