% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{convertCNFtoDNF}
\alias{convertCNFtoDNF}
\title{a function for converting formulas in a CNF form to a DNF form}
\usage{
convertCNFtoDNF(CNFclauses)
}
\description{
a function for converting formulas in a CNF form to a DNF form
}

