% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RuleInduction.OtherFuncCollections.R
\name{findBestCN2Rule}
\alias{findBestCN2Rule}
\title{It is an auxiliary function for greedy covering the data with decision rules using CN2 algorithm.}
\usage{
findBestCN2Rule(dataSet, clsVec, uniqueCls, K = 10)
}
\description{
It is an auxiliary function for greedy covering the data with decision rules using CN2 algorithm.
}

