% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RuleInduction.OtherFuncCollections.R
\name{firstWin}
\alias{firstWin}
\title{the first-win voting strategy}
\usage{
firstWin(object, ruleSet, ...)
}
\arguments{
\item{object}{a class "RuleSetRST"}

\item{a}{data vector representing the object for which predictions are to be made}
}
\value{
a predicted value
}
\description{
An auxiliary function that performs the first-win voting strategy.
}

