% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NearestNeigbour.OtherFuncCollections.R
\name{FRNN.alg}
\alias{FRNN.alg}
\title{The fuzzy-rough nearest neighbour algorithm}
\usage{
FRNN.alg(decision.table, newdata, type.method = "C.FRNN.FRST",
  control = list())
}
\arguments{
\item{decision.table}{a data frame representing the decision table. See \code{\link{BC.IND.relation.FRST}}.}

\item{newdata}{a data frame or matrix (p x n) of data for the test process, where p is the number of instances and
n is the number of conditional attributes (input variables);}

\item{control}{a list of other parameters which depends on two approaches we are going to choose, as follows:
\itemize{
\item implicator/t-norm based model: we should set \code{type.LU = "implicator.tnorm"} and following parameters.
        \itemize{
          \item \code{k}: the number of neighbours. The default value is 5.
          \item \code{type.aggregation}: the type of aggregation.
          \item \code{type.relation}: the type of relation. The default value is \code{c("tolerance", "eq.1")}.
          \item \code{type.implicator}: the type of implicator operator. The default value is "lukasiewicz".
         }
\item vaquely quantified rough set model: for using this type, we should set \code{type.LU = "vqrs"} and following parameters.
         \itemize{
          \item \code{k}:  the number of neighbor. The default value is 5.
          \item \code{q.some}: a numeric of alpha and beta parameter of VQRS. The default value is \code{c(0.1, 0.6)}.
          \item \code{q.most}: a numeric of alpha and beta parameter of VQRS. The default value is \code{c(0.2, 1)}.
          \item \code{type.relation}:  the type of relation. The default value is \code{c("tolerance", "eq.1")}.
          \item \code{type.aggregation}: the type of aggregation.
          }
}
The description of those parameters could be seen at \code{\link{BC.LU.approximation.FRST}}.}
}
\value{
a matrix of predicted classes of newdata.
}
\description{
This is a function that implements the fuzzy-rough nearest neighbour algorithm.
}
\references{
Richard Jensen and Chris Cornelis, "Fuzzy-rough Nearest Neighbour Classificition and Prediction",
Theoretical Computer Science, vol. 412, p. 5871 - 5884 (2011).
}
\seealso{
\code{\link{C.FRNN.O.FRST}},
\code{\link{C.POSNN.FRST}}
}

