% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{FS.DAAR.heuristic.RST}
\alias{FS.DAAR.heuristic.RST}
\title{The DAAR.heuristic}
\usage{
FS.DAAR.heuristic.RST(decision.table, decisionIdx = ncol(decision.table),
  qualityF = X.gini, nAttrs = NULL,
  allowedRandomness = 1/ncol(decision.table),
  nOfProbes = ncol(decision.table), permsWithinINDclasses = FALSE, ...)
}
\arguments{
\item{decision.table}{a \code{"DecisionTable"} class representing the decision table. See \code{\link{SF.asDecisionTable}}.}

\item{decisionIdx}{an integer value representing an index of the decision attribute.}

\item{qualityF}{a function representing the quality of subset of attributes. In this package, the following functions have been included:
\itemize{
\item \code{X.entropy}: See \code{\link{X.entropy}}.
\item \code{X.gini}: See \code{\link{X.gini}}.
\item \code{X.nOfConflicts}: See \code{\link{X.nOfConflicts}}.
\item \code{X.nOfConflictsLog}: See \code{\link{X.nOfConflictsLog}}.
\item \code{X.nOfConflictsSqrt}: See \code{\link{X.nOfConflictsSqrt}}.
}}

\item{nAttrs}{a vector representing indexes of conditional attributes.}

\item{allowedRandomness}{a number of dataset that will be shuffled.}

\item{nOfProbes}{a number of dataset.}

\item{...}{other parameters.}
}
\value{
A class \code{"FeatureSubset"} that contains the following components:
\itemize{
\item \code{reduct}: a list representing a single reduct. In this case, it could be a superreduct or just a subset of features.
\item \code{type.method}: a string representing the type of method which is \code{"greedy.heuristic"}.
\item \code{type.task}: a string showing the type of task which is \code{"feature selection"}.
\item \code{model}: a string representing the type of model. In this case, it is \code{"RST"} which means rough set theory.
\item \code{relevanceProbabilities}: average the probability of attribute relevance using random probes
\item \code{epsilon}: a value between 0 and 1 representing a margin.
}
}
\description{
This is a function used for implementing a DAAR heuristic method for feature selection based on RST.
}
\examples{
###################################################
## Example 1: Evaluate reduct and generate
##            new decision table
###################################################
data(RoughSetData)
decision.table <- RoughSetData$hiring.dt

## evaluate a single reduct
res.1 <- FS.DAAR.heuristic.RST(decision.table)

## generate a new decision table corresponding to the reduct
new.decTable <- SF.applyDecTable(decision.table, res.1)
}
\seealso{
\code{\link{FS.quickreduct.RST}} and \code{\link{FS.reduct.computation}}.
}

