% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{func.conorm}
\alias{func.conorm}
\title{conorm calculation on two variables}
\usage{
func.conorm(right.val, init.val, t.conorm)
}
\arguments{
\item{right.val}{a value of one attribute on right side}

\item{init.val}{a value of attribute}

\item{cotnorm}{a type of t-norm}
}
\description{
This function is used to calculate conorm of two variables
}

