% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NearestNeigbour.OtherFuncCollections.R
\name{get.knearest}
\alias{get.knearest}
\title{This function is to get k nearest neighbor}
\usage{
get.knearest(decision.table, newdata.i, k, method.type)
}
\arguments{
\item{decision.table}{data are used for training}

\item{k}{number of neighbor}

\item{method.type}{a type of method used to calculate distance}

\item{newdata}{data are used for testing}
}
\description{
This function is to get k nearest neighbor
}

