% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{print.FeatureSubset}
\alias{print.FeatureSubset}
\title{The print method of FeatureSubset objects}
\usage{
\method{print}{FeatureSubset}(x, ...)
}
\arguments{
\item{x}{an object inheriting from \code{"FeatureSubset"} class. See \code{\link{FS.reduct.computation}}.}

\item{...}{parameters passes to other functions (currently omitted).}
}
\value{
Prints its argument and returns it invisibly.
}
\description{
This is a print method for FeatureSubset objects.
}
\examples{
###########################################################
## Example : Classification problem
###########################################################
data(RoughSetData)
decision.table <- RoughSetData$hiring.dt

res.1 <- FS.reduct.computation(decision.table)
print(res.1)
}
\author{
Andrzej Janusz
}

