% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{qualityGain}
\alias{qualityGain}
\title{This is an auxiliary function for computing decision reducts}
\usage{
qualityGain(vec, decisionVec, INDclasses, baseChaos, chaosFunction = X.gini)
}
\description{
This is an auxiliary function for computing decision reducts
}

