% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{quickreduct.alg}
\alias{quickreduct.alg}
\title{QuickReduct algorithm based on rough set theory and fuzzy rough set theory}
\usage{
quickreduct.alg(decision.table, type.method = "fuzzy.dependency",
  type.QR = "fuzzy.QR", control = list())
}
\arguments{
\item{decision.table}{a data frame representing decision table. See \code{\link{BC.IND.relation.FRST}}.}

\item{type.method}{a type of the methods}

\item{type.QR}{a type of quick reduct}

\item{control}{a list of other parameters}
}
\value{
reduct.
}
\description{
This is a function for implementing the algorithms of quick reduct for feature selection
based on rough set theory and fuzzy rough set theory.
}
\details{
There are several variety of algorithms based on QuickReduct algorithm (\code{\link{FS.quickreduct.RST}}) and
(\code{\link{FS.quickreduct.FRST}})
It should be noted that new decision table is produced based on one chosen reduct.
}
\seealso{
\code{\link{FS.brute.force.RST}}, \code{\link{FS.heuristic.filter.RST}}
}

