% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{randomize.attrs}
\alias{randomize.attrs}
\title{It is used to randomize attributes}
\usage{
randomize.attrs(attributes)
}
\arguments{
\item{attributes}{a matrix of attributes}
}
\description{
It is used to randomize attributes
}

