% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RuleInduction.OtherFuncCollections.R
\name{rulesVoting}
\alias{rulesVoting}
\title{An auxiliary function that classifies a data object using custom voting strategy.}
\usage{
rulesVoting(object, ruleSet, votingMethod = X.ruleStrength)
}
\description{
An auxiliary function that classifies a data object using custom voting strategy.
}

