\name{X.bestFirst}
\alias{X.bestFirst}
\title{the best-first voting strategy function}
\usage{
X.bestFirst(object, ruleValues, consequents, majorityCls, ...)
}
\arguments{
  \item{object}{a class "RuleSetRST".}

  \item{ruleValues}{values of rules.}

  \item{consequents}{values on the consequent part.}

  \item{majorityCls}{a value representing majority class of
  decision attribute.}

  \item{...}{other parameters.}
}
\value{
predicted values.
}
\description{
It is an auxiliary function of the best-first voting
strategy.
}

