% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{X.entropy}
\alias{X.entropy}
\title{The information gain measure function}
\usage{
X.entropy(decisionDistrib)
}
\arguments{
\item{decisionDistrib}{a distribution of values on the decision attribute.}
}
\value{
\code{qualityF}.
}
\description{
It is an auxiliary function for \code{qualityF} parameter in the \code{\link{FS.greedy.heuristic.reduct.RST}} and \code{\link{FS.greedy.heuristic.superreduct.RST}} functions.
It is based on \emph{entropy} as a measure of information(Shannon, 1948).
}
\references{
C. E. Shannon, "A Mathematical Theory of Communication", Bell System Technical Journal, vol. 27, p. 379 - 423, 623 - 656 (1948).
}

