% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleInduction.OtherFuncCollections.R
\name{X.laplace}
\alias{X.laplace}
\title{Rule voting by the Laplace estimate}
\usage{
X.laplace(rule)
}
\arguments{
\item{rule}{a decision rule, i.e. element of an "RuleSetRST" object}
}
\value{
a numerical weight of the vote
}
\description{
A function returning a weight of rule's vote understood as the Laplace estimate of its confidence.
}
\seealso{
Other currently available voting methods are: \code{\link{X.ruleStrength}}, \code{\link{X.rulesCounting}}.
}
\author{
Andrzej Janusz
}
