\name{X.nOfConflicts}
\alias{X.nOfConflicts}
\title{The discernibility measure function}
\usage{
  X.nOfConflicts(decisionDistrib)
}
\arguments{
  \item{decisionDistrib}{a distribution of values on the
  decision attribute.}
}
\value{
  \code{qualityF}.
}
\description{
  It is an auxiliary function for the \code{qualityF}
  parameter in the
  \code{\link{FS.greedy.heuristic.reduct.RST}} and
  \code{\link{FS.greedy.heuristic.superreduct.RST}}
  functions.
}

