% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{X.nOfConflictsLog}
\alias{X.nOfConflictsLog}
\title{The discernibility measure function based on \code{log2}}
\usage{
X.nOfConflictsLog(decisionDistrib)
}
\arguments{
\item{decisionDistrib}{a distribution of values on the decision attribute.}
}
\value{
\code{qualityF}.
}
\description{
It is an auxiliary function for the \code{qualityF} parameter in the \code{\link{FS.greedy.heuristic.reduct.RST}} and \code{\link{FS.greedy.heuristic.superreduct.RST}} functions.
}

