% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleInduction.OtherFuncCollections.R
\name{X.rulesCounting}
\alias{X.rulesCounting}
\title{Rule voting by counting matching rules}
\usage{
X.rulesCounting(rule)
}
\arguments{
\item{rule}{a decision rule, i.e. element of an "RuleSetRST" object}
}
\value{
a numerical weight of the vote
}
\description{
A function returning an equal vote's weight for every rule. It corresponds to voting by counting the
matching rules.
}
\seealso{
Other currently available voting methods are: \code{\link{X.ruleStrength}}, \code{\link{X.laplace}}.
}
\author{
Andrzej Janusz
}
