% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{as.character.RuleSetRST}
\alias{as.character.RuleSetRST}
\title{The \code{as.character} method for RST rule sets}
\usage{
\method{as.character}{RuleSetRST}(x, ...)
}
\arguments{
\item{x}{a \code{"RuleSetRST"} object. See \code{\link{RI.LEM2Rules.RST}}.}

\item{...}{the other parameters.}
}
\value{
Converts rules from a set into their character representation.
}
\description{
A function for converting a set of rules into their character representation.
}
\examples{
###########################################################
## Example : Converting a set of decision rules
###########################################################
data(RoughSetData)
hiring.data <- RoughSetData$hiring.dt

rules <- RI.LEM2Rules.RST(hiring.data)

as.character(rules)
}
\author{
Andrzej Janusz
}

