% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{print.RuleSetRST}
\alias{print.RuleSetRST}
\title{The print function of rules based on RST}
\usage{
\method{print}{RuleSetRST}(object, ...)
}
\arguments{
\item{object}{a \code{"RuleSetRST"} object. See \code{\link{RI.LEM2Rules.RST}}.}

\item{...}{the other parameters.}
}
\value{
print prints its argument and returns it invisibly
}
\description{
This is a print method for RuleSetRST objects.
}
\examples{
###########################################################
## Example : Classification problem
###########################################################
data(RoughSetData)
decision.table <- RoughSetData$hiring.dt

rules <- RI.LEM2Rules.RST(decision.table)

print(rules)
}

