% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IOFunctions.R
\name{print.RuleSetRST}
\alias{print.RuleSetRST}
\title{The print function for RST rule sets}
\usage{
\method{print}{RuleSetRST}(x, ...)
}
\arguments{
\item{x}{a \code{"RuleSetRST"} object. See \code{\link{RI.LEM2Rules.RST}}.}

\item{...}{the other parameters.}
}
\value{
prints its argument and returns it invisibly
}
\description{
A print method for RuleSetRST objects.
}
\examples{
###########################################################
## Example : Classification problem
###########################################################
data(RoughSetData)
hiring.data <- RoughSetData$hiring.dt

rules <- RI.LEM2Rules.RST(hiring.data)

print(rules)
}
\author{
Andrzej Janusz
}

