% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{X.entropy}
\alias{X.entropy}
\title{The entropy measure}
\usage{
X.entropy(decisionDistrib)
}
\arguments{
\item{decisionDistrib}{an integer vector corresponding to a distribution of attribute values}
}
\value{
a numeric value indicating entropy of an attribute.
}
\description{
An auxiliary function for the \code{qualityF} parameter in the \code{\link{FS.greedy.heuristic.reduct.RST}}, \code{\link{FS.DAAR.heuristic.RST}} and \code{\link{FS.greedy.heuristic.superreduct.RST}} functions.
It is based on \emph{entropy} as a measure of information(Shannon, 1948).
}
\author{
Andrzej Janusz
}
\references{
C. E. Shannon, "A Mathematical Theory of Communication", Bell System Technical Journal, vol. 27, p. 379 - 423, 623 - 656 (1948).
}

