% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FeatureSelection.OtherFuncCollections.R
\name{X.nOfConflicts}
\alias{X.nOfConflicts}
\title{The discernibility measure}
\usage{
X.nOfConflicts(decisionDistrib)
}
\arguments{
\item{decisionDistrib}{an integer vector corresponding to a distribution of decision attribute values}
}
\value{
a numeric value indicating a number of conflicts in a decision attribute
}
\description{
It is an auxiliary function for the \code{qualityF} parameter in the \code{\link{FS.greedy.heuristic.reduct.RST}} and \code{\link{FS.greedy.heuristic.superreduct.RST}} functions.
}

