% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RuleInduction.OtherFuncCollections.R
\name{X.rulesCounting}
\alias{X.rulesCounting}
\title{Rule voting by counting matching rules}
\usage{
X.rulesCounting(rule)
}
\arguments{
\item{a}{decision rule, i.e. element of an "RuleSetRST" object}
}
\value{
weight of the vote
}
\description{
A function returning an equal vote's weight for every rule.
}

