% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RuleInduction.OtherFuncCollections.R
\name{X.ruleStrength}
\alias{X.ruleStrength}
\title{Rule voting by strength of the rule}
\usage{
X.ruleStrength(rule)
}
\arguments{
\item{a}{decision rule, i.e. element of an "RuleSetRST" object}
}
\value{
weight of the vote
}
\description{
A function returning a weight of rule's vote understood as strength of the rule.
It is defined as a product of cardinality of matching training object set and rule length.
}

